C#: Add option to use running server(s) for IT. (#4206)
* Add option to use running server(s) for IT.

+++ csharp/tests/Valkey.Glide.IntegrationTests/TestConfiguration.cs
@@ -15,14 +15,17 @@ public class TestConfiguration : IDisposable
     public static List<(string host, ushort port)> STANDALONE_HOSTS { get; internal set; } = [];
     public static List<(string host, ushort port)> CLUSTER_HOSTS { get; internal set; } = [];
     public static Version SERVER_VERSION { get; internal set; } = new();
+    public static bool TLS { get; internal set; } = false;
 
     public static StandaloneClientConfigurationBuilder DefaultClientConfig() =>
         new StandaloneClientConfigurationBuilder()
-            .WithAddress(STANDALONE_HOSTS[0].host, STANDALONE_HOSTS[0].port);
+            .WithAddress(STANDALONE_HOSTS[0].host, STANDALONE_HOSTS[0].port)
+            .WithTls(TLS);
 
     public static ClusterClientConfigurationBuilder DefaultClusterClientConfig() =>
         new ClusterClientConfigurationBuilder()
-            .WithAddress(CLUSTER_HOSTS[0].host, CLUSTER_HOSTS[0].port);
+            .WithAddress(CLUSTER_HOSTS[0].host, CLUSTER_HOSTS[0].port)
+            .WithTls(TLS);
 
     public static GlideClient DefaultStandaloneClientWithExtraTimeout()
     => GlideClient.CreateClient(DefaultClientConfig().WithRequestTimeout(1000).Build()).GetAwaiter().GetResult();
@@ -120,22 +123,36 @@ public class TestConfiguration : IDisposable
 
         _scriptDir = Path.Combine(projectDir, "..", "utils");
 
-        // Stop all if weren't stopped on previous test run
-        StopServer(false);
+        TLS = Environment.GetEnvironmentVariable("tls") == "true";
 
-        // Delete dirs if stop failed due to https://github.com/valkey-io/valkey-glide/issues/849
-        // Not using `Directory.Exists` before deleting, because another process may delete the dir while IT is running.
-        string clusterLogsDir = Path.Combine(_scriptDir, "clusters");
-        try
+        if (Environment.GetEnvironmentVariable("cluster-endpoints") is { } || Environment.GetEnvironmentVariable("standalone-endpoints") is { })
         {
-            Directory.Delete(clusterLogsDir, true);
+            string? clusterEndpoints = Environment.GetEnvironmentVariable("cluster-endpoints");
+            CLUSTER_HOSTS = clusterEndpoints is null ? [] : ParseHostsString(clusterEndpoints);
+            string? standaloneEndpoints = Environment.GetEnvironmentVariable("standalone-endpoints");
+            STANDALONE_HOSTS = standaloneEndpoints is null ? [] : ParseHostsString(standaloneEndpoints);
+            _startedServer = false;
         }
-        catch (DirectoryNotFoundException) { }
+        else
+        {
+            _startedServer = true;
+            // Stop all if weren't stopped on previous test run
+            StopServer(false);
 
-        // Start cluster
-        CLUSTER_HOSTS = StartServer(true);
-        // Start standalone
-        STANDALONE_HOSTS = StartServer(false);
+            // Delete dirs if stop failed due to https://github.com/valkey-io/valkey-glide/issues/849
+            // Not using `Directory.Exists` before deleting, because another process may delete the dir while IT is running.
+            string clusterLogsDir = Path.Combine(_scriptDir, "clusters");
+            try
+            {
+                Directory.Delete(clusterLogsDir, true);
+            }
+            catch (DirectoryNotFoundException) { }
+
+            // Start cluster
+            CLUSTER_HOSTS = StartServer(true, TLS);
+            // Start standalone
+            STANDALONE_HOSTS = StartServer(false, TLS);
+        }
         // Get redis version
         SERVER_VERSION = GetServerVersion();
 
@@ -149,11 +166,15 @@ public class TestConfiguration : IDisposable
     public void Dispose()
     {
         ResetTestClients();
-        // Stop all
-        StopServer(true);
+        if (_startedServer)
+        {
+            // Stop all
+            StopServer(true);
+        }
     }
 
     private readonly string _scriptDir;
+    private readonly bool _startedServer;
 
     private static void TestConsoleWriteLine(string message) =>
         TestContext.Current.SendDiagnosticMessage(message);
@@ -261,4 +282,7 @@ public class TestConfiguration : IDisposable
         string line = lines.FirstOrDefault(l => l.Contains("valkey_version")) ?? lines.First(l => l.Contains("redis_version"));
         return new(line.Split(':')[1]);
     }
+
+    private static List<(string host, ushort port)> ParseHostsString(string @string)
+        => [.. @string.Split(',').Select(s => s.Split(':')).Select(s => (host: s[0], port: ushort.Parse(s[1])))];
 }
